Added Docs to start with, docker support, and folder structure
The tools for making db migrations and other support are added in, although they will need their docker images upgraded. This is also where we will have the mongodb and the mongo migration tool. The phinx settings will need to be updated and we will use that for the sql lite and the mysql migrations

I converted the 0.4 docs to markdown and split them up in the folders, later on the Doxygen will use them in the outputted documentation, although I suspect the md files will be split further up

Still need issue tracker online to tie into documentation